Separate Event Tests into individual files
