+++ github-activity/src/test/java/com/github/niallantony/IssuesEventTest.java
@@ -1,0 +1,5 @@
+package com.github.niallantony;
+
+public class IssuesEventTest {
+
+}
